Decision Tree
Decision tree is a binary tree that recursively split dataset until we left with pure leaf node.
That is the data with only one type of class. Let’s take whole data in our parent node or root node and based on the condition we will split the data until root node will come.